Change NSBezierPath to use NSInteger and CGFloat and adjust uses to this.
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/gui/trunk@36054 72102866-910b-0410-8b05-ffd578937521

+++ Source/GSThemeTools.m
@@ -411,7 +411,7 @@
 {
   NSBezierPath *p = [NSBezierPath bezierPath];
   NSPoint point;
-  float radius;
+  CGFloat radius;
 
   // make smaller than enclosing frame
   cellFrame = NSInsetRect(cellFrame, 4, floor(cellFrame.size.height * 0.1875));
